+++ quill-cassandra-zio/src/test/scala/io/getquill/context/cassandra/zio/DecodeNullSpec.scala
@@ -12,7 +12,7 @@ class DecodeNullSpec extends ZioCassandraSpec {
       val ret =
         for {
           _ <- testZioDB.run(writeEntities.delete)
-          _ <- testZioDB.run(writeEntities.insertValue(lift(insertValue)))
+          _ <- testZioDB.run(writeEntities.insertValue(lift(insertee)))
           result <- testZioDB.run(query[DecodeNullTestEntity])
         } yield {
           result
@@ -20,7 +20,7 @@ class DecodeNullSpec extends ZioCassandraSpec {
 
       result(ret.foldCause(
         cause => {
-          cause.died must equal(true)
+          cause.isDie must equal(true)
           cause.dieOption match {
             case Some(e: Exception) =>
               e.isInstanceOf[IllegalStateException] must equal(true)
@@ -39,5 +39,5 @@ class DecodeNullSpec extends ZioCassandraSpec {
 
   case class DecodeNullTestWriteEntity(id: Int, value: Option[Int])
 
-  val insertValue = DecodeNullTestWriteEntity(0, None)
+  val insertee = DecodeNullTestWriteEntity(0, None)
 }