+++ quill-cassandra-zio/src/test/scala/io/getquill/context/cassandra/zio/examples/other/ExampleApp.scala
@@ -1,0 +1,25 @@
+package io.getquill.context.cassandra.zio.examples.other
+
+import io.getquill._
+import zio.Console.printLine
+import zio.ZIOAppDefault
+
+object ExampleApp extends ZIOAppDefault {
+
+  object MyZioPostgresContext extends CassandraZioContext(Literal)
+  import MyZioPostgresContext.*
+
+  case class Person(name: String, age: Int)
+
+  val zioSessionLayer =
+    CassandraZioSession.fromPrefix("testStreamDB")
+
+  override def run = {
+    inline def people = quote {
+      query[Person]
+    }
+    MyZioPostgresContext.run(people)
+      .tap(result => printLine(result.toString))
+      .provide(zioSessionLayer).exitCode
+  }
+}