+++ quill-jasync/src/main/scala/io/getquill/context/jasync/Decoders.scala
@@ -9,7 +9,6 @@ import com.github.jasync.sql.db.RowData
 import io.getquill.MappedEncoding
 
 import io.getquill.util.Messages.fail
-import org.joda.time.{ DateTime => JodaDateTime, LocalDate => JodaLocalDate, LocalTime => JodaLocalTime, LocalDateTime => JodaLocalDateTime }
 
 import scala.reflect.{ ClassTag, classTag }
 
@@ -40,7 +39,7 @@ trait Decoders {
           case value if f.isDefinedAt(value) => f(value)
           case value =>
             fail(
-              s"Value '$value' at index $index can't be decoded to '${classTag[T].runtimeClass}'"
+              s"Value '$value' at index $index can't be decoded to '${classTag[T].runtimeClass}' (value is: ${value.getClass})"
             )
         }
       }
@@ -133,38 +132,20 @@ trait Decoders {
 
   implicit val byteArrayDecoder: Decoder[Array[Byte]] = decoder[Array[Byte]](PartialFunction.empty, SqlTypes.TINYINT)
 
-  implicit val jodaDateTimeDecoder: Decoder[JodaDateTime] = decoder[JodaDateTime]({
-    case dateTime: JodaDateTime           => dateTime
-    case localDateTime: JodaLocalDateTime => localDateTime.toDateTime
-  }, SqlTypes.TIMESTAMP)
-
-  implicit val jodaLocalDateDecoder: Decoder[JodaLocalDate] = decoder[JodaLocalDate]({
-    case localDate: JodaLocalDate => localDate
-  }, SqlTypes.DATE)
-
-  implicit val jodaLocalDateTimeDecoder: Decoder[JodaLocalDateTime] = decoder[JodaLocalDateTime]({
-    case localDateTime: JodaLocalDateTime => localDateTime
-  }, SqlTypes.TIMESTAMP)
-
   implicit val dateDecoder: Decoder[Date] = decoder[Date]({
-    case localDateTime: JodaLocalDateTime => localDateTime.toDate
-    case localDate: JodaLocalDate         => localDate.toDate
+    case date: LocalDateTime  => Date.from(date.atZone(dateTimeZone).toInstant)
+    case date: LocalDate      => Date.from(date.atStartOfDay.atZone(dateTimeZone).toInstant)
+    case date: OffsetDateTime => Date.from(date.toInstant)
   }, SqlTypes.TIMESTAMP)
 
-  implicit val decodeZonedDateTime: MappedEncoding[JodaDateTime, ZonedDateTime] =
-    MappedEncoding(jdt => ZonedDateTime.ofInstant(Instant.ofEpochMilli(jdt.getMillis), ZoneId.of(jdt.getZone.getID)))
-
-  implicit val decodeOffsetDateTime: MappedEncoding[JodaDateTime, OffsetDateTime] =
-    MappedEncoding(jdt => OffsetDateTime.ofInstant(Instant.ofEpochMilli(jdt.getMillis), ZoneId.of(jdt.getZone.getID)))
-
-  implicit val decodeLocalDate: MappedEncoding[JodaLocalDate, LocalDate] =
-    MappedEncoding(jld => LocalDate.of(jld.getYear, jld.getMonthOfYear, jld.getDayOfMonth))
-
-  implicit val decodeLocalTime: MappedEncoding[JodaLocalTime, LocalTime] =
-    MappedEncoding(jlt => LocalTime.of(jlt.getHourOfDay, jlt.getMinuteOfHour, jlt.getSecondOfMinute))
-
-  implicit val decodeLocalDateTime: MappedEncoding[JodaLocalDateTime, LocalDateTime] =
-    MappedEncoding(jldt => LocalDateTime.ofInstant(jldt.toDate.toInstant, ZoneId.systemDefault()))
-
-  implicit val localDateDecoder: Decoder[LocalDate] = mappedDecoder(decodeLocalDate, jodaLocalDateDecoder)
+  implicit val localDateDecoder: Decoder[LocalDate] = decoder[LocalDate](PartialFunction.empty, SqlTypes.DATE)
+  implicit val localDateTimeDecoder: Decoder[LocalDateTime] = decoder[LocalDateTime](PartialFunction.empty, SqlTypes.TIMESTAMP)
+  implicit val zonedDateTimeDecoder: Decoder[ZonedDateTime] = decoder[ZonedDateTime]({
+    // Use the currently-configured timezone to translate the DB date back (as opposed to the zone on the timezone column itself)
+    // to the ZonedDateTime since that information is lost during encoding
+    case date: LocalDateTime  => ZonedDateTime.from(date.atZone(dateTimeZone).toInstant)
+    case date: LocalDate      => ZonedDateTime.from(date.atStartOfDay.atZone(dateTimeZone).toInstant)
+    // If we have an OffsetDateTime returned, assume that it is in the timezone of the system
+    case date: OffsetDateTime => date.atZoneSameInstant(dateTimeZone)
+  }, SqlTypes.TIMESTAMP)
 }