+++ quill-jdbc-zio/src/test/scala/io/getquill/sqlite/PrepareJdbcSpec.scala
@@ -1,16 +1,18 @@
 package io.getquill.sqlite
 
+import io.getquill.context.qzio.ImplicitSyntax.Implicit
+import javax.sql.DataSource
 import java.sql.{ Connection, ResultSet }
 import io.getquill.PrepareZioJdbcSpecBase
-import io.getquill.Prefix
+
 import org.scalatest.BeforeAndAfter
 import io.getquill._
 
 class PrepareJdbcSpec extends PrepareZioJdbcSpecBase with BeforeAndAfter {
 
-  def prefix = Prefix("testSqliteDB")
-  val context: testContext.type = testContext
-  import testContext._
+  implicit val ds: Implicit[DataSource] = Implicit(pool)
+  val context: testContext.underlying.type = testContext.underlying
+  import testContext.underlying._
 
   before {
     testContext.run(query[Product].delete).runSyncUnsafe()