+++ quill-jdbc-zio/src/test/scala/io/getquill/sqlserver/sqlserver.scala
@@ -1,5 +1,10 @@
 package io.getquill
 
+import io.getquill.context.qzio.ImplicitSyntax.Implicit
+import io.getquill.ZioSpec.runLayerUnsafe
+import io.getquill.jdbczio.Quill
+
 package object sqlserver {
-  object testContext extends SqlServerZioJdbcContext(Literal) with TestEntities
+  val pool = runLayerUnsafe(Quill.DataSource.fromPrefix("testSqlServerDB"))
+  object testContext extends Quill.SqlServer(Literal, pool) with TestEntities
 }