+++ quill-jdbc/src/test/scala/io/getquill/context/jdbc/oracle/ScalarValueSpec.scala
@@ -17,6 +17,6 @@
 //   }
 
 //   "Multi Scalar Select with Infix" in {
-//     context.run("foo" + infix"""'bar'""".as[String]) mustEqual "foobar"
+//     context.run("foo" + sql"""'bar'""".as[String]) mustEqual "foobar"
 //   }
 // }