+++ quill-sql/src/main/scala/io/getquill/SqlMirrorContext.scala
@@ -7,7 +7,7 @@ import io.getquill.context.AstSplicing
 import io.getquill.context.mirror.MirrorSession
 
 /** Workaround for IntelliJ SCL-20185. Inheriting MirrorContextBase directly so that `run` methods have autocomplete. */
-class SqlMirrorContext[Idiom <: BaseIdiom, Naming <: NamingStrategy](val idiom: Idiom, val naming: Naming)
+class SqlMirrorContext[+Idiom <: BaseIdiom, +Naming <: NamingStrategy](val idiom: Idiom, val naming: Naming)
     extends MirrorContextBase[Idiom, Naming]
     with AstSplicing
     with SqlContext[Idiom, Naming]